requirements.txt
streamlit==1.28.1
tavily-python==0.3.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas
plotly